requirements.txt
ldap3
omero-certificates
omero-py
toml
zeroc-ice
